Still playing with post excerpt on single.php
There are multiple versions of single.php saved with various names they
are all in various states of disarray. None of them work. The closes
one is the current single.php file however the excerpt still does not
show. Attempted to use WP_Query to no avail. Will play with this
further